+++ tests/AzureAISearchSimulator.Api.Tests/Services/IndexServiceScoringProfileValidationTests.cs
@@ -1,6 +1,7 @@
 using AzureAISearchSimulator.Api.Services;
 using AzureAISearchSimulator.Core.Configuration;
 using AzureAISearchSimulator.Core.Models;
+using AzureAISearchSimulator.Search;
 using AzureAISearchSimulator.Storage.Repositories;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
@@ -21,7 +22,9 @@ public class IndexServiceScoringProfileValidationTests
         var repositoryMock = new Mock<IIndexRepository>();
         var settings = Options.Create(new SimulatorSettings());
         var logger = new Mock<ILogger<IndexService>>();
-        _sut = new IndexService(repositoryMock.Object, settings, logger.Object);
+        var luceneSettings = Options.Create(new LuceneSettings { IndexPath = Path.Combine(Path.GetTempPath(), "sim-test-" + Guid.NewGuid().ToString("N")) });
+        var luceneManager = new LuceneIndexManager(new Mock<ILogger<LuceneIndexManager>>().Object, luceneSettings);
+        _sut = new IndexService(repositoryMock.Object, settings, luceneManager, logger.Object);
     }
 
     private static SearchIndex CreateIndexWithFields() => new()